requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
icmplib==3.0.4
idna==3.10
iperf3==0.1.11
json-fix==0.5.2
markdown-it-py==3.0.0
mdurl==0.1.2
nest-asyncio==1.6.0
netifaces==0.11.0
platformdirs==4.3.6
prometheus_client==0.21.1
psutil==6.1.0
py-directus==0.0.30
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-magic==0.4.27
rich==13.9.4
sniffio==1.3.1
speedtest-cli==2.1.3
starlette==0.41.3
typing_extensions==4.12.2
unicorn==2.1.1
uvicorn==0.32.1
websockets==14.1
